Remove -s from makepkg
The -s was kept aroung because the package base `python-virtualfish`
failed to build without it. I first blamed this on some aur rpc bug
because it was missing some deps that were listed in python-virtualfish.
As is turned out python-virtualfish actually does some things wrong in
it's package build and if it was formated correctly this wouldnt be
a problem.

I kept the -s in just so it would build even though it might have had
some side affects. makepkg not respecting the --dbpath for example.

From now on python-virtualfish will fail to build when you dont have all
the deps installed. This is their fault and will not be fixed here.