Remove test for formatKeysToImport()
The recent changes to formatting broke this test. While it would be easy
to go back and fix it, I personaly do not like tests that rely on things
such as formatting or an internet connection. This is why I have removed
the test, I believe the other test for key importing should be more that
sufficient.