Fix -Sc sometimes cleaning split packages
If a split package was installed and there was no package that matched
the name of the package base Yay would remove it even though there could
be other packages installed under that base but with a different name.

Now only clean a package if there are no installed packages under the
packagebase.