Implement c29a80 properly
It seems the pkgbase is null for installed packages that are not part of
a split package. It was priviously assumed that if a package was not
part of a split package, pkgbase == pkgname would always be true.

Instead try to use pkgbase and if it does not exist fall back to
pkgname.