Allow running yay via sudo
This commit allows running yay via sudo. It does not however let you run
yay as root. Instead the SUDO_UID and SUDO_GID are used to sudo back
into the previous user that originally called yay.

This could allow for `PACMAN=yay makepkg` as a way to build packages
with makepkg while having dependency resolution for the AUR.

The return code problem would also need to be solved for this to work.